Fix typing and add upgrades

+++ bot/utils.py
@@ -10,8 +10,10 @@ from bot.user import User
 if t.TYPE_CHECKING:
     from bot.model import Model
 
+
 class Utils:
     """A class containing utility functions for the bot."""
+
     def __init__(self, model) -> None:
         self.model: Model = model
 
@@ -29,9 +31,8 @@ class Utils:
         if char_id not in user_character_ids:
             await ctx.respond(f"**{character.first_name} {character.last_name}** is not in your list!")
             return None
-        
-        return character
 
+        return character
 
     async def search_characters(
         self,
@@ -41,6 +42,9 @@ class Utils:
         limit: int = 0,
         fuzzy: bool = False,
     ) -> list[Character]:
+        if self.model.dbpool is None:
+            return []
+
         async with self.model.dbpool.acquire() as conn:
             fn = None
             ln = None
@@ -49,7 +53,7 @@ class Utils:
                 ln = None
                 names = name.split(" ")
                 if len(names) > 1:
-                    fn = " ".join(names[0 : len(names) - 1])
+                    fn = " ".join(names[0: len(names) - 1])
                     ln = names[len(names) - 1]
 
             sql = "SELECT * FROM characters "
@@ -57,7 +61,7 @@ class Utils:
 
             if id:
                 sql += f"AND ID = $1 "
-                args.append(id)
+                args.append(str(id))
             if fn:
                 sql += f"AND LOWER(first_name) = LOWER($2) "
                 fn = "%" + fn + "%" if fuzzy else fn
@@ -79,10 +83,14 @@ class Utils:
             if limit > 0:
                 sql += f"LIMIT {limit}"
 
-            if "$3" not in sql: sql = sql.replace("$4","$3")
-            if "$2" not in sql: sql = sql.replace("$3","$2")
-            if "$1" not in sql: sql = sql.replace("$2","$1")
-            if "$2" not in sql: sql = sql.replace("$3","$2")
+            if "$3" not in sql:
+                sql = sql.replace("$4", "$3")
+            if "$2" not in sql:
+                sql = sql.replace("$3", "$2")
+            if "$1" not in sql:
+                sql = sql.replace("$2", "$1")
+            if "$2" not in sql:
+                sql = sql.replace("$3", "$2")
 
             characters_a = await conn.fetch(sql, *args)
 
@@ -105,13 +113,17 @@ class Utils:
             return character_list
 
     async def add_characters_to_db(self) -> None:
+        if self.model.dbpool is None:
+            return
+
         async with self.model.dbpool.acquire() as conn:
             f = open("bot/data/db.csv", "r", encoding="utf8")
             reader = csv.reader(f, delimiter="|")
             next(reader)
             data = []
             for x in reader:
-                data.append([int(x[0]), x[1], x[2], x[3], x[4], int(x[5]), x[6], x[7]])
+                data.append([int(x[0]), x[1], x[2], x[3],
+                            x[4], int(x[5]), x[6], x[7]])
 
             await conn.execute("DROP TABLE IF EXISTS characters")
             await conn.execute(
